Moved MT5 Password Manager to its own component file
